+++ src/emit.rs
@@ -5,6 +5,7 @@ use crate::type_expr::{
     NativeTypeInfo,
     ObjectField,
     TypeArray,
+    TypeDefinition,
     TypeExpr,
     TypeInfo,
     TypeIntersection,
@@ -140,17 +141,59 @@ impl<'ctx> EmitCtx<'ctx> {
     }
 }
 
+struct SepList<'a, T>(&'a [T], &'static str);
+
+impl<'a, T> Emit for SepList<'a, T>
+where
+    T: Emit,
+{
+    fn emit(&self, ctx: &mut EmitCtx<'_>) -> io::Result<()> {
+        let Self(elements, separator) = self;
+        let mut first = true;
+        for element in *elements {
+            if !first {
+                write!(ctx.w, "{}", separator)?;
+            }
+            element.emit(ctx)?;
+            first = false;
+        }
+        Ok(())
+    }
+}
+
+struct Generics<'a, T>(&'a [T]);
+
+impl<'a, T> Emit for Generics<'a, T>
+where
+    T: Emit,
+{
+    fn emit(&self, ctx: &mut EmitCtx<'_>) -> io::Result<()> {
+        let Self(args) = self;
+        if !args.is_empty() {
+            write!(ctx.w, "<")?;
+            SepList(args, ",").emit(ctx)?;
+            write!(ctx.w, ">")?;
+        }
+        Ok(())
+    }
+}
+
 impl Emit for TypeExpr {
     fn emit(&self, ctx: &mut EmitCtx<'_>) -> io::Result<()> {
         match self {
-            TypeExpr::Ref(TypeInfo::Native(NativeTypeInfo { def })) => {
-                def.emit(ctx)
+            TypeExpr::Ref(TypeInfo::Native(NativeTypeInfo { r#ref })) => {
+                r#ref.emit(ctx)
             },
             TypeExpr::Ref(TypeInfo::Defined(DefinedTypeInfo {
-                docs: _,
-                path,
-                name,
-                def: _,
+                def:
+                    TypeDefinition {
+                        docs: _,
+                        path,
+                        name,
+                        generic_vars: _,
+                        def: _,
+                    },
+                generic_args,
             })) => {
                 write!(ctx.w, "{}.", ctx.options.root_namespace)?;
                 for path_part in *path {
@@ -158,6 +201,7 @@ impl Emit for TypeExpr {
                     write!(ctx.w, ".")?;
                 }
                 name.emit(ctx)?;
+                Generics(generic_args).emit(ctx)?;
                 Ok(())
             },
             TypeExpr::Name(type_name) => type_name.emit(ctx),
@@ -178,25 +222,14 @@ impl Emit for TypeName {
         let Self {
             path,
             name,
-            generics,
+            generic_args,
         } = self;
         for path_part in *path {
             path_part.emit(ctx)?;
             write!(ctx.w, ".")?;
         }
         name.emit(ctx)?;
-        if !generics.is_empty() {
-            write!(ctx.w, "<")?;
-            let mut first = true;
-            for generic in *generics {
-                if !first {
-                    write!(ctx.w, ",")?;
-                }
-                generic.emit(ctx)?;
-                first = false;
-            }
-            write!(ctx.w, ">")?;
-        }
+        Generics(generic_args).emit(ctx)?;
         Ok(())
     }
 }
@@ -215,14 +248,7 @@ impl Emit for TypeTuple {
         let Self { docs, elements } = self;
         docs.emit(ctx)?;
         write!(ctx.w, "[")?;
-        let mut first = true;
-        for element in *elements {
-            if !first {
-                write!(ctx.w, ",")?;
-            }
-            element.emit(ctx)?;
-            first = false;
-        }
+        SepList(elements, ",").emit(ctx)?;
         write!(ctx.w, "]")?;
         Ok(())
     }
@@ -273,14 +299,7 @@ impl Emit for TypeUnion {
             write!(ctx.w, "never")?;
         } else {
             write!(ctx.w, "(")?;
-            let mut first = true;
-            for part in *members {
-                if !first {
-                    write!(ctx.w, "|")?;
-                }
-                part.emit(ctx)?;
-                first = false;
-            }
+            SepList(members, "|").emit(ctx)?;
             write!(ctx.w, ")")?;
         }
         Ok(())
@@ -295,14 +314,7 @@ impl Emit for TypeIntersection {
             write!(ctx.w, "any")?;
         } else {
             write!(ctx.w, "(")?;
-            let mut first = true;
-            for part in *members {
-                if !first {
-                    write!(ctx.w, "&")?;
-                }
-                part.emit(ctx)?;
-                first = false;
-            }
+            SepList(members, "&").emit(ctx)?;
             write!(ctx.w, ")")?;
         }
         Ok(())
@@ -354,42 +366,32 @@ where
 
 impl EmitCtx<'_> {
     fn emit_type(&mut self, info: &'static TypeInfo) -> io::Result<()> {
-        for dep_info in info.iter_refs() {
-            self.emit_def(dep_info)?;
-        }
-        Ok(())
-    }
-
-    fn emit_def(&mut self, info: DefinedTypeInfo) -> io::Result<()> {
-        let DefinedTypeInfo {
+        for TypeDefinition {
             docs,
             path,
             name,
+            generic_vars,
             def,
-        } = info;
-        self.stats.type_definitions += 1;
-        docs.emit(self)?;
-        if !path.is_empty() {
-            write!(self.w, "export namespace ")?;
-            let mut first = true;
-            for path_part in path {
-                if !first {
-                    write!(self.w, ".")?;
-                }
-                path_part.emit(self)?;
-                first = false;
+        } in crate::iter_def_deps::IterDefDeps::new(info)
+        {
+            self.stats.type_definitions += 1;
+            docs.emit(self)?;
+            if !path.is_empty() {
+                write!(self.w, "export namespace ")?;
+                SepList(path, ".").emit(self)?;
+                write!(self.w, "{{")?;
             }
-            write!(self.w, "{{")?;
+            write!(self.w, "export type ")?;
+            name.emit(self)?;
+            Generics(generic_vars).emit(self)?;
+            write!(self.w, "=")?;
+            def.emit(self)?;
+            write!(self.w, ";")?;
+            if !path.is_empty() {
+                write!(self.w, "}}")?;
+            }
+            writeln!(self.w)?;
         }
-        write!(self.w, "export type ")?;
-        name.emit(self)?;
-        write!(self.w, "=")?;
-        def.emit(self)?;
-        write!(self.w, ";")?;
-        if !path.is_empty() {
-            write!(self.w, "}}")?;
-        }
-        writeln!(self.w)?;
         Ok(())
     }
 }