+++ src/lib.rs
@@ -120,12 +120,14 @@
 
 mod emit;
 mod impls;
-mod iter_refs;
+mod iter_def_deps;
 pub mod type_expr;
 
-pub use crate::{
-    emit::{write_definition_file, DefinitionFileOptions, Stats, TypeDef},
-    impls::Blob,
+pub use crate::emit::{
+    write_definition_file,
+    DefinitionFileOptions,
+    Stats,
+    TypeDef,
 };
 
 /// A derive proc-macro for the [`TypeDef`] trait.